Работаем с датой.Изменение2
