+++ bootstrap.php
@@ -3,5 +3,10 @@
 date_default_timezone_set('Europe/Moscow');
 
 require_once 'functions/template.php';
+require_once 'functions/validators.php';
+require_once 'functions/db.php';
 
-require_once 'functions/helpers.php';
+error_reporting(E_ALL);
+
+ini_set("display_error", true);
+ini_set("error_reporting", E_ALL);
