+++ functions/template.php
@@ -1,4 +1,145 @@
 <?php
+
+/**
+ * Возвращает корректную форму множественного числа
+ * Ограничения: только для целых чисел
+ *
+ * Пример использования:
+ * $remaining_minutes = 5;
+ * echo "Я поставил таймер на {$remaining_minutes} " .
+ *     get_noun_plural_form(
+ *         $remaining_minutes,
+ *         'минута',
+ *         'минуты',
+ *         'минут'
+ *     );
+ * Результат: "Я поставил таймер на 5 минут"
+ *
+ * @param int $number Число, по которому вычисляем форму множественного числа
+ * @param string $one Форма единственного числа: яблоко, час, минута
+ * @param string $two Форма множественного числа для 2, 3, 4: яблока, часа, минуты
+ * @param string $many Форма множественного числа для остальных чисел
+ *
+ * @return string Рассчитанная форма множественнго числа
+ */
+function get_noun_plural_form(int $number, string $one, string $two, string $many): string
+{
+    $number = (int)$number;
+    $mod10 = $number % 10;
+    $mod100 = $number % 100;
+
+    switch (true) {
+        case ($mod100 >= 11 && $mod100 <= 20):
+            return $many;
+
+        case ($mod10 > 5):
+            return $many;
+
+        case ($mod10 === 1):
+            return $one;
+
+        case ($mod10 >= 2 && $mod10 <= 4):
+            return $two;
+
+        default:
+            return $many;
+    }
+}
+
+/**
+ * Подключает шаблон, передает туда данные и возвращает итоговый HTML контент
+ * @param string $name Путь к файлу шаблона относительно папки templates
+ * @param array $data Ассоциативный массив с данными для шаблона
+ * @return string Итоговый HTML
+ */
+function include_template($name, array $data = [])
+{
+    $name = 'templates/' . $name;
+    $result = '';
+
+    if (!is_readable($name)) {
+        return $result;
+    }
+
+    ob_start();
+    extract($data);
+    require $name;
+
+    $result = ob_get_clean();
+
+    return $result;
+}
+
+/**
+ * Возвращает img-тег с обложкой видео для вставки на страницу
+ * @param string $youtube_url Ссылка на youtube видео
+ * @return string
+ */
+function embed_youtube_cover($youtube_url)
+{
+    $res = "";
+    $id = extract_youtube_id($youtube_url);
+
+    if ($id) {
+        $src = sprintf("https://img.youtube.com/vi/%s/mqdefault.jpg", $id);
+        $res = '<img alt="youtube cover" width="320" height="120" src="' . $src . '" />';
+    }
+
+    return $res;
+}
+
+/**
+ * Извлекает из ссылки на youtube видео его уникальный ID
+ * @param string $youtube_url Ссылка на youtube видео
+ * @return array
+ */
+function extract_youtube_id($youtube_url)
+{
+    $id = false;
+
+    $parts = parse_url($youtube_url);
+
+    if ($parts) {
+        if ($parts['path'] == '/watch') {
+            parse_str($parts['query'], $vars);
+            $id = $vars['v'] ?? null;
+        } else {
+            if ($parts['host'] == 'youtu.be') {
+                $id = substr($parts['path'], 1);
+            }
+        }
+    }
+
+    return $id;
+}
+
+/**
+ * @param $index
+ * @return false|string
+ */
+function generate_random_date($index)
+{
+    $deltas = [['minutes' => 59], ['hours' => 23], ['days' => 6], ['weeks' => 4], ['months' => 11]];
+    $dcnt = count($deltas);
+
+    if ($index < 0) {
+        $index = 0;
+    }
+
+    if ($index >= $dcnt) {
+        $index = $dcnt - 1;
+    }
+
+    $delta = $deltas[$index];
+    $timeval = rand(1, current($delta));
+    $timename = key($delta);
+
+    $ts = strtotime("$timeval $timename ago");
+    $dt = date('Y-m-d H:i:s', $ts);
+
+    return $dt;
+}
+
 /**
  * Обрезает строку
  *
@@ -40,7 +181,7 @@ function dateDifferent(string $date) : string
 {
     $markPost = strtotime($date);
     $markNow = time();
-    $markDiff = $markNow - $markPost;
+    $markDiff =  $markNow - $markPost;
 
     $datePost = date_create($date);
     $dateNow = date_create("now");
@@ -48,19 +189,19 @@ function dateDifferent(string $date) : string
 
     if ($markDiff/60 < 60) {
         //если до текущего времени прошло меньше 60 минут, то формат будет вида “% минут назад”;
-        $numDiff = ceil($markDiff/60);
+        $numDiff = round($markDiff/60);
         $interval = "$numDiff ".get_noun_plural_form($numDiff, "минута", "минуты", "минут")." назад";
     } elseif ($markDiff/3600 < 24) {
         //если до текущего времени прошло больше 60 минут, но меньше 24 часов, то формат будет вида “% часов назад”;
-        $numDiff = ceil($markDiff/3600);
+        $numDiff = round($markDiff/3600);
         $interval = "$numDiff ".get_noun_plural_form($numDiff, "час", "часа", "часов")." назад";
     } elseif ($markDiff/86400 < 7) {
         //если до текущего времени прошло больше 24 часов, но меньше 7 дней, то формат будет вида “% дней назад”;
-        $numDiff = ceil($markDiff/86400);
+        $numDiff = round($markDiff/86400);
         $interval = "$numDiff ".get_noun_plural_form($numDiff, "день", "дня", "дней")." назад";
     } elseif ($markDiff/(86400*7) < 5) {
         //если до текущего времени прошло больше 7 дней, но меньше 5 недель, то формат будет вида “% недель назад”;
-        $numDiff = ceil($markDiff/(86400*7));
+        $numDiff = round($markDiff/(86400*7));
         $interval = "$numDiff ".get_noun_plural_form($numDiff, "неделя", "недели", "недель")." назад";
     } else {
         //если до текущего времени прошло больше 5 недель, то формат будет вида “% месяцев назад”.
