+++ index.php
@@ -46,6 +46,6 @@ $cards = [
 
 $page_content = include_template("main.php", $cards);
 
-$layout_content = include_template("layout.php", ['content' => $page_content, 'title' => 'readme: популярное', 'user' => $user_name]);
+$layout_content = include_template("layout.php", ['content' => $page_content, 'title' => 'readme: популярное', 'user' => $user_name, 'is_auth' => $is_auth]);
 
 print($layout_content);
