+++ templates/layout.php
@@ -35,7 +35,7 @@
             </div>
         </form>
         <div class="header__nav-wrapper">
-            <?php if ($is_auth === 1) : ?>
+            <?php if ($data['is_auth'] === 1) : ?>
             <nav class="header__nav">
                 <ul class="header__my-nav">
                     <li class="header__my-page header__my-page--popular">
